Remove cssnano and use node sass compressed style
