+++ package.json
@@ -33,7 +33,6 @@
     "chalk": "^1.1.3",
     "chokidar": "^1.6.1",
     "cross-spawn": "^4.0.2",
-    "cssnano": "^3.8.0",
     "glob-promise": "^1.0.6",
     "mkdirp": "^0.5.1",
     "node-sass": "^3.10.1",
